requirements.txt
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.28.0
gTTS>=2.2.3
SpeechRecognition>=3.8.1
wikipedia>=1.4.0
PyAudio>=0.2.11